+++ csharp/HelloWorld.cs
@@ -1,25 +1,13 @@
-using System;
-namespace HelloWorld
+namespace Sample
 {
-    class HelloWorld
+    public class HelloWorld
     {
-        static void Main()
-        {
-            Console.WriteLine("HelloWorld");
-            int a=10,b=20;
-            int c=plus(a,b);
-            Console.WriteLine("plus={0}",c);
-            HelloWorld hello = new HelloWorld();
-            int d=hello.minus(a,b);
-            Console.WriteLine("minus={0}",d);
-        }
-
-        static int plus(int a,int b)
+        public static int plus(int a,int b)
         {
             return a+b;
         }
 
-        int minus(int a,int b)
+        public int minus(int a,int b)
         {
             return a-b;
         }
