+++ csharp/README.md
@@ -1,2 +1,3 @@
-$ mcs HelloWorld.cs
+$ mcs /target:library /out:HelloWorld.dll HelloWorld.cs
+$ mcs -r:HelloWorld.dll Hello.cs
 $ mono HelloWorld.exe
